v1.0.1: fix SIZE_T typedef wrong size on x64 (prevents ACCESS_VIOLATION on newer Windows)
Found while debugging Mica regression after Windows 11 25H2 update KB5089549.

Bug:
  ffi.cdef declared `typedef unsigned long SIZE_T;`
  unsigned long is 4 bytes on Windows (LLP64), but SIZE_T (= ULONG_PTR) is
  pointer-sized — 8 bytes on x64. This made WINDOWCOMPOSITIONATTRIBDATA
  4 bytes too small, so ulDataSize was read as garbage by Windows.

Symptom history:
  Older Windows builds tolerated the size mismatch silently.
  Post KB5089549 (May 2026, Windows 11 25H2), validation got stricter and
  any call to SetWindowCompositionAttribute with a bad struct now crashes
  the process with ACCESS_VIOLATION (0xC0000005).

Fix:
  typedef uintptr_t SIZE_T;
  Same pattern as the v1.0.0 ULONG_PTR fix. LuaJIT FFI resolves uintptr_t
  to the platform-correct pointer width.

Note: KB5089549 also appears to disable rendering of WCA_ACCENT_POLICY blur
effects entirely, so even with the correct struct the plugin currently
produces no visible blur. Same fate as MicaForEveryone, StartAllBack, DSX
on this Windows build. This fix is forward-compatible — if Microsoft
restores the rendering in a future update, the plugin resumes working
without further changes.

+++ backend/main.lua
@@ -22,7 +22,12 @@ typedef LONG_PTR LPARAM;
 typedef unsigned int UINT;
 typedef long HRESULT;
 typedef int INT;
-typedef unsigned long SIZE_T;
+// Kitsune fix #2: SIZE_T is pointer-sized (= ULONG_PTR) — 8 bytes on x64, not 4.
+// Upstream declared it as `unsigned long` (4 bytes) which made WINDOWCOMPOSITIONATTRIBDATA
+// 4 bytes too small. Older Windows versions tolerated the mismatch silently; newer ones
+// (post KB5089549, May 2026) validate stricter and crash the process with ACCESS_VIOLATION
+// (0xC0000005) when ulDataSize is read as garbage.
+typedef uintptr_t SIZE_T;
 typedef char CHAR;
 typedef CHAR *LPSTR;
